requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
qrcode[pil]==7.4.2
fpdf2==2.7.6